Authorization

+++ internal/domain/user.go
@@ -3,7 +3,7 @@ package domain
 import "time"
 
 type User struct {
-	ID         int    `json:"-" db:"id"`
+	ID         int64  `json:"-" db:"id"`
 	Email      string `json:"email" binding:"required"`
 	Username   string `json:"username" binding:"required"`
 	Password   string `json:"password" binding:"required"`
